chore(core): move config directory out of utils
