+++ packages/core/src/config/config.ts
