+++ packages/core/src/config/index.ts
